+++ includes/class-cocart-session.php
@@ -112,8 +112,35 @@ class CoCart_API_Session {
 			$override_cart   = true;  // Override the cart by default.
 			$notify_customer = false; // Don't notify the customer by default.
 			$redirect        = false; // Don't safely redirect the customer to the cart after loading by default.
+			$user_id         = 0;     // Guest user unless stated later.
 
 			wc_nocache_headers();
+
+			// Check the user is logged in. If true a different cart cannot be loaded so just return.
+			if ( is_user_logged_in() ) {
+				$current_user = wp_get_current_user();
+				$user_id      = $current_user->ID;
+
+				// Compare the user ID with the cart key.
+				if ( $user_id == $cart_key ) {
+					CoCart_Logger::log( sprintf( __( 'Cart key "%s" is already loaded as the user is logged in.', 'cart-rest-api-for-woocommerce' ), $cart_key ), 'errro' );
+				} else {
+					CoCart_Logger::log( sprintf( __( 'Customer is already logged in. Cart key "%s" cannot be loaded into session.', 'cart-rest-api-for-woocommerce' ), $cart_key ), 'error' );
+				}
+				return;
+			}
+			// If user is not logged in, check that the cart key does not belong to a user registered.
+			else {
+				$user = get_user_by( 'id', $cart_key );
+
+				// If the user exists then just return.
+				if ( ! is_null( $user->ID ) ) {
+					CoCart_Logger::log( __( 'Cart key is recognised as a registered user on site. Cannot be loaded into session.', 'cart-rest-api-for-woocommerce' ), 'error' );
+					return;
+				}
+			}
+
+			// At this point, the cart should load into session with no issues as we have passed verification.
 
 			// Check if we are keeping the cart currently set via the web.
 			if ( ! empty( $_REQUEST['keep-cart'] ) && is_bool( $_REQUEST['keep-cart'] ) !== true ) {